+++ src/index.js
@@ -2,8 +2,8 @@ import Form from "./Forms/Form";
 import Input from "./Forms/Input";
 import TextArea from "./Forms/TextArea";
 import CheckBox from "./Forms/CheckBox";
-import Select from "./Forms/SelectionBox/Select";
-import Option from "./Forms/SelectionBox/Option";
+import Select from "./Forms/Select";
+import Option from "./Forms/Option";
 import ValidationTypes from "./Forms/ValidationTypes";
 
 export { Form, Input, TextArea, CheckBox, ValidationTypes, Select, Option };
